Downloadable product links - add new data provider in separated module. #142
