+++ src/module-vsbridge-downloadable/composer.json
@@ -1,0 +1,21 @@
+{
+    "name": "divante/module-vsbridge-downloadable",
+    "type": "magento2-module",
+    "authors":[
+        {
+            "name":"Agata",
+            "email":"afirlejczyk@divante.pl"
+        }
+    ],
+    "require": {
+        "elasticsearch/elasticsearch": "~5.0"
+    },
+    "autoload": {
+        "files": [
+            "registration.php"
+        ],
+        "psr-4": {
+            "Divante\\VsbridgeDownloadable\\": ""
+        }
+    }
+}
