+++ src/module-vsbridge-downloadable/etc/module.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
+    <module name="Divante_VsbridgeDownloadable" setup_version="1.2.0">
+        <sequence>
+            <module name="Divante_VsbridgeIndexerCatalog"/>
+        </sequence>
+    </module>
+</config>
