+++ src/module-vsbridge-downloadable/etc/vsbridge_indices.xml
@@ -1,0 +1,11 @@
+<indices xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:noNamespaceSchemaLocation="urn:magento:module:Divante_VsbridgeIndexerCore:etc/vsbridge_indices.xsd">
+
+    <index identifier="vue_storefront_catalog">
+        <type name="product" mapping="Divante\VsbridgeIndexerCatalog\Index\Mapping\Product">
+            <data_providers>
+                <data_provider name="downloadable_data">Divante\VsbridgeDownloadable\Model\Indexer\DataProvider\Product\DownloadableData</data_provider>
+            </data_providers>
+        </type>
+    </index>
+</indices>
