+++ src/module-vsbridge-downloadable/registration.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * Copyright Divante Sp. z o.o.
+ * See LICENSE_DIVANTE.txt for license details.
+ */
+
+declare(strict_types=1);
+
+use \Magento\Framework\Component\ComponentRegistrar;
+
+ComponentRegistrar::register(
+    ComponentRegistrar::MODULE,
+    'Divante_VsbridgeDownloadable',
+    __DIR__
+);
